Stall machanism was changed,added dhd
The mechanism for stalling was updated for something less clumsy.
After some checks,the new mechanism works just fine, but there are
still some bugs to be fixed (listed at the head of the document).
Also, a Data Hazard Detection (dhd) was added.It is still under
testing, so it's commmented out